+++ srcs/window/map/vertex/get_vertex.c
@@ -6,7 +6,7 @@
 /*   By: rbroque <rbroque@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/09 14:10:09 by rbroque           #+#    #+#             */
-/*   Updated: 2023/01/25 11:35:00 by rbroque          ###   ########.fr       */
+/*   Updated: 2023/01/26 11:13:48 by rbroque          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,10 +33,10 @@ void	set_vertex(t_vertex *vertex,
 	get_color_comp(&vertex->color);
 }
 
-void	get_vertex(t_map *map, const size_t x, const size_t y, const char *format)
+void	get_vertex(t_map *map, const int x, const int y, const char *format)
 {
-	const int	new_x = (int)x;
-	const int	new_y = (int)y;
+	const int	new_x = x;
+	const int	new_y = y;
 	const int	new_z = get_altitude(format, &map->minz, &map->maxz);
 
 	set_vertex(&map->vertex[y][x], new_x, new_y, new_z, WHITE);
